requirements.txt
faster-whisper>=1.1.0
sentence-transformers>=3.3.0
transformers>=4.40.0
torch>=2.1.0
faiss-cpu>=1.7.0
wikipedia-api>=0.6.0
flask>=3.0.0
flask-socketio>=5.4.0
gevent>=24.0.0
gevent-websocket>=0.10.0
numpy>=1.26.0
